changed the cost data type again to best practice for monetary numbers - BigDecimal.

+++ wistlish-backend/src/main/java/com/example/wistlish_app/models/Item.java
@@ -3,6 +3,7 @@ package com.example.wistlish_app.models;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -18,7 +19,7 @@ public class Item {
     @Column(length = 2083) // URL length limit
     private String imageUrl;
     private int quantity;
-    private double cost;
+    private BigDecimal cost;
     @Column(columnDefinition = "TINYINT") // Using TINYINT for boolean in MySQL
     private boolean isClaimed;
     private LocalDate createdOn;
@@ -30,7 +31,7 @@ public class Item {
 
     // Constructors
     public Item() {}
-    public Item(String name, String itemUrl, String imageUrl, int quantity, double cost, Wishlist list) {
+    public Item(String name, String itemUrl, String imageUrl, int quantity, BigDecimal cost, Wishlist list) {
         this.name = name;
         this.itemUrl = itemUrl;
         this.imageUrl = imageUrl;
@@ -78,11 +79,11 @@ public class Item {
         this.quantity = quantity;
     }
 
-    public double getCost() {
+    public BigDecimal getCost() {
         return cost;
     }
 
-    public void setCost(double cost) {
+    public void setCost(BigDecimal cost) {
         this.cost = cost;
     }
